# Simplify RebuildIndex