--- a/go/libraries/doltcore/sqle/writer/noms_table_writer_test.go
+++ b/go/libraries/doltcore/sqle/writer/noms_table_writer_test.go
@@ -162,7 +162,7 @@ func TestTableEditor(t *testing.T) {
 
 			ctx := sqle.NewTestSQLCtx(context.Background())
 			root, _ := dEnv.WorkingRoot(context.Background())
-			opts := editor.Options{Deaf: dEnv.DbEaFactory()}
+			opts := editor.Options{Deaf: dEnv.DbEaFactory(), Tempdir: dEnv.TempTableFilesDir()}
 			db := sqle.NewDatabase("dolt", dEnv.DbData(), opts)
 			err := dsess.DSessFromSess(ctx.Session).AddDB(ctx, getDbState(t, db, dEnv))
 			require.NoError(t, err)